move ellip_from_axis_ratio2epsilon from sl_hammocks_pipeline.py to param_util.py
